+++ plone/app/z3cform/widget.py
@@ -137,7 +137,7 @@ class DateWidget(BaseWidget, HTMLInputWidget):
     _converter = DateWidgetConverter
     _formater = 'date'
 
-    pattern = 'pickadate'
+    pattern = 'date-picker'
     pattern_options = BaseWidget.pattern_options.copy()
 
     def _base_args(self):
@@ -210,6 +210,7 @@ class DatetimeWidget(DateWidget, HTMLInputWidget):
     _converter = DatetimeWidgetConverter
     _formater = 'dateTime'
 
+    pattern = 'datetime-picker'
     pattern_options = DateWidget.pattern_options.copy()
 
     default_timezone = None